feat: move _utils to helpers
